requirements.txt
langchain_google_genai
langchain_core
streamlit~=1.35.0
python-dotenv~=1.0.1